requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
EasyProcess==1.1
h11==0.13.0
idna==3.3
outcome==1.1.0
Pillow==9.0.0
pycparser==2.21
pygame==2.1.2
pyOpenSSL==22.0.0
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
PyVirtualDisplay==2.2
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
